Add new component to handle fetch Missions and add test to new component
