+++ src/Components/Navbar.js
@@ -1,18 +1,10 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
-import { useDispatch, useSelector } from 'react-redux';
-import { getMissions } from '../redux/missions/missions';
+import MissionsLink from './MissionsLink';
 import icon from '../img/icon.png';
 import '../css/Navbar.css';
 
 export default function Navbar() {
-  const dispatch = useDispatch();
-  const stateM = useSelector((state) => state.missionsStore.missions);
-  const missionClickHandler = () => {
-    if (!stateM.length) {
-      dispatch(getMissions());
-    }
-  };
   return (
     <nav className="flex">
       <div className="logo flex">
@@ -30,14 +22,7 @@ export default function Navbar() {
           </NavLink>
         </li>
         <li>
-          <NavLink
-            onClick={missionClickHandler}
-            className="link"
-            to="/missions"
-            style={({ isActive }) => (isActive ? { textDecoration: 'underline' } : {})}
-          >
-            Missions
-          </NavLink>
+          <MissionsLink />
         </li>
         <li>
           <NavLink
